Implement Login endpoint

+++ web-server/src/main/java/org/maciejmarczak/financemanager/operation/OperationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-interface OperationRepository extends JpaRepository<Operation, Long> {
-    List<Operation> getAllByIssuerId(Long issuerId);
+interface OperationRepository extends JpaRepository<Operation, Integer> {
+    List<Operation> getAllByIssuerId(Integer issuerId);
 }
