+++ web-server/src/main/java/org/maciejmarczak/financemanager/user/UserRepository.java
@@ -2,6 +2,6 @@ package org.maciejmarczak.financemanager.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-interface UserRepository extends JpaRepository<User, Long> {
+interface UserRepository extends JpaRepository<User, Integer> {
     User findByEmail(String email);
 }
